Finish moving graphwidget and mapwidget to its own file
